ReviewController updated
Major changes while getting reviews. I implemented TextSearch Places API to search a place by its name. Then there is a separate method to get the reviews. I get as a response in Insomnia a cutomized json object with all the data we need. Next step is to join the ReviewController and MapsController.